+++ lib/klaviyo_api/collections/event_collection.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+module KlaviyoAPI::Collections
+  class EventCollection < MarkerCollection
+    def next_page_marker_name
+      'since'
+    end
+  end
+end